src/test/java/dev/gothickit/zenkit/world/vob/TriggerWorldStartTest.java
package dev.gothickit.zenkit.world.vob;

import dev.gothickit.zenkit.GameVersion;
import dev.gothickit.zenkit.LogLevel;
import dev.gothickit.zenkit.Logger;
import dev.gothickit.zenkit.Util;
import dev.gothickit.zenkit.capi.ZenKit;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

class TriggerWorldStartTest {
	@BeforeAll
	static void beforeAll() {
		ZenKit.load();
		Logger.setDefault(LogLevel.DEBUG);
	}

	@Test
	void load() {
		var vob = new TriggerWorldStart(Util.getResource("G2/VOb/zCTriggerWorldStart.zen"), GameVersion.GOTHIC_2);
		assertEquals("EVT_TROLL_GRAVE_MOVER_01", vob.getTarget());
		assertTrue(vob.getFireOnce());
	}
}